+++ authorization-service/serverless.yml
@@ -24,9 +24,9 @@ provider:
   #  SAMPLE_ENV_VAR: ${env:SAMPLE_ENV_VAR}
 
 functions:
-  hello:
-    handler: handler.hello
-    events:
-      - http:
-          path: hello
-          method: get
+  basicAuthorizer:
+    handler: handler.basicAuthorizer
+
+custom:
+  bundle:
+    linting: false
